+++ Typin.Tests/Data/Commands/Valid/WithEnvironmentVariablesCommand.cs
@@ -7,10 +7,10 @@
     [Command("cmd")]
     public class WithEnvironmentVariablesCommand : SelfSerializeCommandBase
     {
-        [CommandOption("opt-a", 'a', EnvironmentVariableName = "ENV_OPT_A")]
+        [CommandOption("opt-a", 'a', FallbackVariableName = "ENV_OPT_A")]
         public string? OptA { get; set; }
 
-        [CommandOption("opt-b", 'b', EnvironmentVariableName = "ENV_OPT_B")]
+        [CommandOption("opt-b", 'b', FallbackVariableName = "ENV_OPT_B")]
         public IReadOnlyList<string>? OptB { get; set; }
     }
 }